requirements.txt
beautifulsoup4==4.12.2
# selenium==4.15.2 # Removed conflicting version
requests==2.31.0
pymongo==4.6.0
python-dotenv==1.0.0
notion-client==2.2.1
schedule==1.2.1
pandas==2.1.4
nltk==3.8.1
python-dateutil==2.8.2
webdriver-manager==4.0.1
mongoengine==0.27.0
selenium==4.16.0
pytz
googletrans-py==4.0.0
scikit-learn==1.3.2
joblib==1.3.2 